backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
numpy>=1.26.4
faiss-cpu==1.9.0.post1
sentence-transformers
scikit-learn
requests
python-dotenv
pandas
pyarrow
rank-bm25==0.2.2
python-dateutil==2.9.0.post0